* move Entities to Entities namespace * implement SelfRegistry class * make CardType and Effect inherit from SelfRegistry
